fix(android): resolve 17 Kotlin compile errors in Top 1-3 code
CI build 29305646089 surfaced 17 compile errors in code that was only
static-checked before (user said "don't compile"). All fixes are
targeted, no behavior change:

CustomImageModel.kt (9 errors): `caps.X ?: providerDefault.X`
was inferred as `Boolean?` because caps.X is `Boolean?`; added explicit
`?: false` so the elvis chain returns non-null `Boolean` as
ImageModelDefinition requires.

UrlSafety.kt (5 errors): hex literals > Int.MAX_VALUE
(0xA9FE0000, 0xAC100000, 0xC0A80000, 0xE0000000, 0xF0000000) were
inferred as Long; added `.toInt()` to coerce to Int (bit pattern
preserved, inCidr compares correctly).

ZoomableImageDialog.kt (1 error): missing
`import androidx.compose.foundation.layout.fillMaxWidth`.

PromptTemplateViewModel.kt (1 error): repo.observeAll() returns
`Flow<Pair<List<Cat>, List<Tpl>>>` but stateIn expected `TemplateData`;
added `.map { (cats, tpls) -> TemplateData(cats, tpls) }` and
`import kotlinx.coroutines.flow.map`.

WorkbenchScreen.kt (1 error): `Icons.AutoMirrored.Filled.List` used
fully-qualified path that collided with kotlin.collections.List;
imported with alias `as AutoMirroredList` and used the short name.

+++ android/app/src/main/java/com/gptimage/playground/data/repository/UrlSafety.kt
@@ -143,10 +143,10 @@ object UrlSafety {
             inCidr(0x0A000000, 8) ||
             inCidr(0x64400000, 10) ||
             inCidr(0x7F000000, 8) ||
-            inCidr(0xA9FE0000, 16) ||
-            inCidr(0xAC100000, 12) ||
-            inCidr(0xC0A80000, 16) ||
-            inCidr(0xE0000000, 4) ||
-            inCidr(0xF0000000, 4)
+            inCidr(0xA9FE0000.toInt(), 16) ||
+            inCidr(0xAC100000.toInt(), 12) ||
+            inCidr(0xC0A80000.toInt(), 16) ||
+            inCidr(0xE0000000.toInt(), 4) ||
+            inCidr(0xF0000000.toInt(), 4)
     }
 }
